move 'RequestedAppointmentListGroup' and unit test to 'pages' directory
